Always derive Default
If a struct derives Serialize but not Default, which schemafy does when a
struct contains a required field, inject a derive for Default.  This
allows using Ignition config structs without manually specifying all the
fields, but it also generates defaults for mandatory fields, which is not
idiomatic.  Generating explicit new() methods would be better, but also
more work.