Avoid Eq derive warnings on clippy 1.63.0+
schemafy never derives Eq, only PartialEq, and clippy now warns about
this.  If we derived Eq ourselves, we could end up having to remove the
derive in a future schema version if a floating-point value is ever added
to the spec.  We don't need Eq for now, so just disable the warning to
keep things simple.

https://rust-lang.github.io/rust-clippy/master/index.html#derive_partial_eq_without_eq